+++ crates/glaredb_http/src/gcs/filesystem.rs
@@ -142,7 +142,7 @@ where
             Some(v) => v
                 .to_str()
                 .context("Failed convert Content-Length header to string")?
-                .parse::<usize>()
+                .parse::<u64>()
                 .context("Failed to parse Content-Length")?,
             None => return Err(DbError::new("Missing Content-Length header for file")),
         };
@@ -269,7 +269,7 @@ where
         self.handle.url.as_str()
     }
 
-    fn size(&self) -> usize {
+    fn size(&self) -> u64 {
         self.handle.len
     }
